requirements.txt
Pillow==10.0.1
djangorestframework-simplejwt==5.3.0
Django==4.2.5
djangorestframework==3.14.0
drf-spectacular==0.26.5




